requirements.txt
Django==3.2
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
